django_openhim_mediators/requirements.txt
requests == 2.18.4
https://github.com/de-laz/openhim-mediator-utils-py.git
djangorestframework == 3.10.3
